Guess Number
In this game, we first declare three variables: num which will hold the randomly generated number, guess which will hold the player's guess, and attempts which will keep track of how many guesses the player has made.

We then use the srand() function to seed the random number generator using the current time so that the numbers generated are different each time the program is run. We then use rand() to generate a random number between 1 and 100 and store it in the num variable.

We then ask the player to guess the number and use a do-while loop to keep prompting the player to guess until they get the correct answer. In each iteration of the loop, we prompt the player to enter their guess, read it in using scanf(), and increment the attempts counter.

We then use if statements to check if the guess is too high or too low and provide feedback to the player. If the guess is correct, we exit the loop and print a congratulatory message along with the number of attempts it took the player to guess the number.