Random Game in Python
In this version of the game, we use the random module to generate a random number between 1 and 100 using the randint() function and store it in the num variable.

We then use a while loop to keep prompting the player to guess until they get the correct answer. In each iteration of the loop, we prompt the player to enter their guess using the input() function and convert it to an integer using the int() function. We also increment the attempts counter.

We then use if statements to check if the guess is too high or too low and provide feedback to the player. If the guess is correct, we exit the loop using the break statement and print a congratulatory message along with the number of attempts it took the player to guess the number.