Arduino motor control.rust
This is a simple code  that uses the Rust arduino_uno library to control the two motors. The portd module is used to configure the digital pins PD3 and PD5 as output pins. The loop function is used to control the direction of the motors. In this example, the motors are set to rotate forward, then stop, then rotate in reverse, and then stop again. The delay_ms function is used to pause the program execution for a certain duration in milliseconds.