Mood Detection in C programming
Facial landmarks are then extracted using a pre-trained shape predictor model. From the facial landmarks, features such as eye aspect ratio (EAR) and mouth aspect ratio (MAR) are computed, which are used as inputs to a support vector machine (SVM) classifier model to predict the mood of the individual. The SVM model was trained on a dataset of labeled facial expressions, which are classified into four mood categories: happy, sad, angry, and neutral. Once the mood is classified, the system displays the predicted mood label on the video feed. This project can be useful in applications such as mental health monitoring, customer service, and entertainment.