Eaglle Scare Crow Algorithm--Bird Robot
This code sets up the four servo motors using the WiringPi library, and then enters an infinite loop where it alternates between flapping the wings, moving the head, and moving the tail. The flapping, head, and tail angles are set in the FLAP_ANGLE, HEAD_ANGLE, and TAIL_ANGLE constants respectively, and the frequencies are set in the FLAP_FREQ, HEAD_FREQ, and TAIL_FREQ constants.