C Random Imagges
This program generates a random image consisting of dots and stars using a 2D array. It first sets the width and height of the image, initializes a random seed using the srand() function, and then generates a random number between 0 and 1 for each pixel in the image. A value of 0 corresponds to a dot, while a value of 1 corresponds to a star.

After generating the random image, the program prints it to the console by iterating over the 2D array and printing a dot or star character for each pixel. The output image will be different each time the program is run due to the random nature of the image generation.