Create Esp32 Hotspot.ino
This code configures the ESP32 in Access Point (AP) mode and sets up a hotspot with the given SSID and password. The WiFi.softAP() function is used to create the hotspot, and the WiFi.softAPIP() function is used to get the IP address of the access point.

Once the ESP32 is in AP mode and the hotspot is configured, the loop() function does nothing. You can upload this code to your ESP32 board and it will start broadcasting a hotspot with the specified SSID and password. You can then connect to the hotspot using any device (such as a phone or laptop) and access the internet through the ESP32.