Compressing Files in C
The program takes two command-line arguments: the name of the source file and the name of the destination file. The program opens the source file for reading and the destination file for writing. It then reads the source file in chunks of 16KB and compresses each chunk using the zlib library. The compressed data is written to the destination file.

Note that this is just a simple example, and there are many ways to compress files in C depending on your requirements.