Create Language Select.c
This program prompts the user to select a language by printing a menu of options, reads the user's input using fgets, removes the newline character from the input using strcspn, and then uses strcmp to compare the input to each language option. If the input matches one of the options, the program prints a message indicating the selected language and then proceeds to implement the functionality for that language. If the input does not match any of the options, the program prints an error message and exits with an error code of 1.