Create Memory Partition.c
In this example code, three memory blocks are allocated using malloc() with sizes of 10, 20, and 30 integers. The allocated memory blocks are then used to store some data, and finally, the memory is freed using the free() function. Note that when using malloc(), it is important to check if the allocation was successful by checking if the returned pointer is NULL.