Create Flush ESP.py
In this code, you first set the path to the esptool.py utility using the esptool_path variable. Then, you set the serial port of the ESP32 board using the serial_port variable. Finally, you set the path to the binary file that you want to flash using the binary_file_path variable.

To flash the binary file to the ESP32 board, you call the subprocess.run() function and pass it a list of arguments. The arguments specify the parameters for the esptool.py utility, such as the chip type (--chip esp32), the serial port (--port), the baud rate (--baud), the flash mode (--flash_mode), the flash frequency (--flash_freq), and the flash size (--flash_size). The last two arguments specify the address (0x1000) where you want to write the binary file to and the path to the binary file itself.

Note that you'll need to replace the values of esptool_path, serial_port, and binary_file_path with the appropriate values for your system and setup.